code/ai-toolkit-requirements.txt
absl-py==2.4.0
accelerate==1.13.0
aiofiles==24.1.0
albucore==0.0.16
albumentations==1.4.15
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
attrs==25.4.0
av==16.0.1
bitsandbytes==0.49.2
brotli==1.2.0
certifi==2026.2.25
charset-normalizer==3.4.4
clean-fid==0.1.35
click==8.3.1
clip-anytorch==2.6.0
contourpy==1.3.3
controlnet_aux==0.0.10
cuda-bindings==12.9.4
cuda-pathfinder==1.4.0
cycler==0.12.1
dctorch==0.1.2
diffusers @ git+https://github.com/huggingface/diffusers@8600b4c10d67b0ce200f664204358747bd53c775
einops==0.8.2
eval_type_backport==0.3.1
fastapi==0.135.1
ffmpy==1.0.0
filelock==3.25.0
flash_attn @ https://github.com/mjun0812/flash-attention-prebuild-wheels/releases/download/v0.7.16/flash_attn-2.8.3%2Bcu128torch2.10-cp311-cp311-linux_x86_64.whl#sha256=7700af158271dc1076386f9cc1188bdb1acf6e12396e8b163c3b8ca77dc2edaf
flatten-json==0.1.14
fonttools==4.61.1
fsspec==2026.2.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.46
gradio==6.8.0
gradio_client==2.2.0
groovy==0.1.2
grpcio==1.78.0
h11==0.16.0
hf-xet==1.3.2
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
invisible-watermark==0.2.0
Jinja2==3.1.6
jsonmerge==1.9.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
k-diffusion==0.1.1.post1
kiwisolver==1.4.9
kornia==0.8.2
kornia_rs==0.1.10
lazy_loader==0.4
lpips==0.1.4
lycoris_lora==1.8.3
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
ninja==1.13.0
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
open_clip_torch==3.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
optimum-quanto==0.2.4
orjson==3.11.7
oyaml==1.0
packaging @ file:///home/task_176104885106445/conda-bld/packaging_1761049078006/work
pandas==3.0.1
peft==0.18.1
pillow==12.1.1
platformdirs==4.9.4
prodigyopt==1.1.2
protobuf==6.33.5
psutil==7.2.2
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
python-slugify==8.0.4
pytorch-fid==0.3.0
pytorch-wavelets==1.3.0
pytz==2026.1.post1
PyWavelets==1.9.0
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-image==0.26.0
scipy==1.12.0
semantic-version==2.10.0
sentencepiece==0.2.1
sentry-sdk==2.54.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
starlette==0.52.1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
text-unidecode==1.3
tifffile==2026.3.3
timm==1.0.25
tokenizers==0.22.2
toml==0.10.2
tomlkit==0.13.3
torch==2.10.0
torchao==0.10.0
torchaudio==2.10.0
torchcodec==0.10.0
torchdiffeq==0.2.5
torchsde==0.2.6
torchvision==0.25.0
tqdm==4.67.3
trampoline==0.1.2
transformers==4.57.3
triton==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
wandb==0.25.0
wcwidth==0.6.0
Werkzeug==3.1.6
zipp==3.23.0
